Move cache to Resources Manager
