+++ WebViewControl/ResourcesManager.cs
@@ -9,6 +9,28 @@ namespace WebViewControl {
 
     public static class ResourcesManager {
 
+        private static readonly AssemblyCache cache = new AssemblyCache();
+
+        public class Resource {
+            public string Extension;
+            public Stream ResourceStream;
+
+            public Resource(string extension, Stream resourceStream) {
+                Extension = extension;
+                ResourceStream = resourceStream;
+            }
+        }
+
+        public static Resource LoadEmbeddedResource(Uri url) {
+            var resourceAssembly = cache.ResolveResourceAssembly(url);
+            var resourcePath = ResourceUrl.GetEmbeddedResourcePath(url);
+
+            var extension = Path.GetExtension(resourcePath.Last()).ToLower();
+            var resourceStream = TryGetResourceWithFullPath(resourceAssembly, resourcePath);
+
+            return new Resource(extension, resourceStream);
+        }
+
         private static Stream InternalTryGetResource(string assemblyName, string defaultNamespace, IEnumerable<string> resourcePath, bool failOnMissingResource) {
             var assembly = AppDomain.CurrentDomain.GetAssemblies().FirstOrDefault(a => a.GetName().Name == assemblyName);
             if (assembly == null) {
